feat: jwt 토큰 커스터마이징 포기.
