+++ backend/album/views.py
@@ -3,7 +3,7 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.views import APIView
-from common.util import user_token_to_data
+from common.utils import user_token_to_data
 from .serializers import *
 
 class AlbumView(APIView):
