+++ backend/common/urls.py
@@ -9,4 +9,5 @@ urlpatterns = [
     path('login/', LoginAPIView.as_view(), name='login'),
     path('signup/', SignupAPIView.as_view(), name='signup'),
     path('logout/', LogoutAPIView.as_view(), name='logout'),
+    # path('api/token/', MyTokenObtainPairView.as_view(), name='token_obtain_pair'),
 ]
